# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Hubx REST API
  version: 0.1.0
  description: >
    Contrato REST inicial da Etapa 2.
    Escopo atual cobre apenas endpoints REST/DRF e APIs HTTP de pagamentos.
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/accounts/accounts/{id}/activate/:
    post:
      tags: [accounts]
      summary: Ativar usuário da organização
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário ativado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStatusResponse"
        "400":
          description: Erro de domínio (já ativo ou payload inválido)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Sem permissão para gerir o usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/accounts/accounts/{id}/deactivate/:
    post:
      tags: [accounts]
      summary: Desativar usuário da organização
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário desativado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStatusResponse"
        "400":
          description: Erro de domínio (já inativo ou payload inválido)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Sem permissão para gerir o usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/accounts/accounts/confirm-email/:
    post:
      tags: [accounts]
      summary: Confirmar e-mail por token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: E-mail confirmado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessageResponse"
        "400":
          description: Token ausente/inválido/expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/accounts/accounts/resend-confirmation/:
    post:
      tags: [accounts]
      summary: Reenviar e-mail de confirmação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        "204":
          description: Processado sem vazamento de existência da conta
        "400":
          description: E-mail ausente ou conta já ativada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/accounts/accounts/request-password-reset/:
    post:
      tags: [accounts]
      summary: Solicitar redefinição de senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        "204":
          description: Processado sem vazamento de existência da conta
        "400":
          description: E-mail ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/accounts/accounts/reset-password/:
    post:
      tags: [accounts]
      summary: Redefinir senha com token de recuperação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Senha redefinida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessageResponse"
        "400":
          description: Dados incompletos, token inválido ou senha inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sessionid
...
  schemas:
    AccountStatusResponse:
      type: object
      required: [detail, id, is_active]
      properties:
        detail:
          type: string
        id:
          type: integer
        is_active:
          type: boolean
    ApiErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
    ApiMessageResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
security:
  - sessionAuth: []